desloppify: split MainViewModel.AttachSidebar + extract Palette partial
AttachSidebar was a 50-line grab-bag. Split into 4 named helpers:

- WireOverview         — OverviewViewModel + focus/back signals
- WireSidebarCallbacks — SpawnSessionRequested + CloseWorktreeSessionsAsync
- WireStoreChangedToStatusBar — store-change → RaiseStatusBarChanged
- HookStatusBarSources (existing)

AttachSidebar is now a 4-line composition that reads as a sequence
of named concerns.

Also pull the command-palette concern out of MainViewModel.cs into
MainViewModel.Palette.cs partial — OpenCommandPaletteAsync and
BuildPaletteActions move together; same Type and namespace.
